Changed Id property type from int to long(for all entities in DB)

+++ ApplicationService/Contracts/ICategoryManagementService.cs
@@ -5,9 +5,9 @@ namespace ApplicationService.Contracts
     public interface ICategoryManagementService
     {
         Task<IEnumerable<CategoryDTO>> GetAsync(string query);
-        Task<CategoryDTO> GetByIdAsync(int id);
+        Task<CategoryDTO> GetByIdAsync(long id);
         Task<CategoryDTO> SaveAsync(CategoryDTO categoryDTO);
         Task<CategoryDTO> UpdateAsync(CategoryDTO categoryDTO);
-        Task DeleteAsync(int id);
+        Task DeleteAsync(long id);
     }
 }
