+++ Repository/Implementations/Repository.cs
@@ -28,7 +28,7 @@ namespace Repository.Implementations
                                                                                                         .Where(expression);
         public async Task<List<T>> FindByConditionAsync<T>(Expression<Func<T, bool>> expression) where T : class => await DbSet<T>()
                                                                                                         .Where(expression).ToListAsync();
-        public async Task<T> FindByIdAsync<T>(int id) where T : class => await DbSet<T>().FindAsync(id);
+        public async Task<T> FindByIdAsync<T>(long id) where T : class => await DbSet<T>().FindAsync(id);
         public async Task CreateAsync<T>(T entity) where T : class => await DbSet<T>().AddAsync(entity);
 
         public void Update<T>(T entity) where T : class => DbSet<T>().Update(entity);
